requirements.txt
# Core library for Google Gemini API integration
google-generativeai